refactor(quality): address #585-588 — coverage, complexity, module split, docs
- language_validators: use `...` for abstract get_language_name body so it is
  not counted as uncovered (restores 100% coverage) (#585)
- reduce cyclomatic complexity of the three worst C-rated functions (#586):
    - summarise `_markdown_body` C(16)->B(9) via `_render_category_group` /
      `_render_flat_files`
    - merge `_apply_diff` C(16)->B(8) via `_git_apply` / `_apply_reject` /
      `_decode_stderr`
    - sync `_clone_template` C(15)->A(4) via `_validate_clone_config` /
      `_resolve_bundle_names` / `_raise_unknown_profile`
- split summarise.py (692 LOC) into a package: `_gather` (data), `_render`
  (formatting), `__init__` (orchestration + back-compat re-exports) (#587)
- move REPOSITORY_ANALYSIS.md into docs/development/ (#588)

All gates green: fmt, ty + mypy --strict, interrogate 100%, deptry, security,
validate, 655 tests passing at 100% coverage.
