+++ src/rhiza/commands/summarise/_render.py
@@ -1,0 +1,364 @@
+"""Rendering helpers for the ``summarise`` command.
+
+Formats gathered change data into markdown, plain text, JSON, or a custom
+Jinja2 template.  Beyond reading a user-supplied Jinja2 template file, this
+module performs no git or filesystem inspection — see
+:mod:`rhiza.commands.summarise._gather` for data collection.
+"""
+
+import json as _json
+from dataclasses import dataclass, field
+from datetime import datetime
+from pathlib import Path
+from typing import Any
+
+import jinja2
+
+from ._gather import _TemplateInfo
+
+
+@dataclass(kw_only=True)
+class SummariseOptions:
+    """Options controlling the output of :func:`generate_pr_description`.
+
+    All fields are keyword-only and default to the standard behaviour so
+    callers only need to set the fields they want to override.
+    """
+
+    include_header: bool = True
+    """Whether to include the header section (markdown / plain formats)."""
+
+    include_footer: bool = True
+    """Whether to include the footer section (markdown / plain formats)."""
+
+    include_categories: bool = True
+    """Whether to group changes by category; when ``False`` a flat list is shown."""
+
+    output_format: str = "markdown"
+    """Output format: ``"markdown"`` (default), ``"plain"``, or ``"json"``."""
+
+    title: str | None = None
+    """Override the section heading; ``None`` uses the built-in default."""
+
+    compare_ref: str | None = None
+    """Compare against this git ref instead of the staged index."""
+
+    jinja2_template: Path | None = field(default=None)
+    """Path to a Jinja2 template file for fully custom output."""
+
+
+def _format_file_list(files: list[str], status_emoji: str) -> list[str]:
+    """Format a list of files with the given status emoji.
+
+    Args:
+        files: List of file paths
+        status_emoji: Emoji to use (✅ for added, 📝 for modified, ❌ for deleted)
+
+    Returns:
+        List of formatted lines
+    """
+    lines = []
+    for f in sorted(files):
+        lines.append(f"- {status_emoji} `{f}`")
+    return lines
+
+
+def _add_category_section(lines: list[str], title: str, count: int, files: list[str], emoji: str) -> None:
+    """Add a collapsible section for a category and change type.
+
+    Args:
+        lines: List to append lines to
+        title: Section title (e.g., "Added", "Modified")
+        count: Number of files
+        files: List of file paths
+        emoji: Status emoji
+    """
+    if not files:
+        return
+
+    lines.append("<details>")
+    lines.append(f"<summary>{title} ({count})</summary>")
+    lines.append("")
+    lines.extend(_format_file_list(files, emoji))
+    lines.append("")
+    lines.append("</details>")
+    lines.append("")
+
+
+def _build_header(template_repo: str, title: str | None = None) -> list[str]:
+    """Build the PR description header.
+
+    Args:
+        template_repo: Template repository name
+        title: Optional override for the section heading
+
+    Returns:
+        List of header lines
+    """
+    header_title = title if title else "## 🔄 Template Synchronization"
+    lines = [header_title, ""]
+    if template_repo:
+        url = f"https://github.com/{template_repo}"
+        repo_link = f"[{template_repo}]({url})"
+        sync_line = f"This PR synchronizes the repository with the {repo_link} template."
+        lines.append(sync_line)
+    else:
+        lines.append("This PR synchronizes the repository with the upstream template.")
+    lines.append("")
+    return lines
+
+
+def _build_summary(changes: dict[str, list[str]]) -> list[str]:
+    """Build the change summary section.
+
+    Args:
+        changes: Dictionary of changes by type
+
+    Returns:
+        List of summary lines
+    """
+    return [
+        "### 📊 Change Summary",
+        "",
+        f"- **{len(changes['added'])}** files added",
+        f"- **{len(changes['modified'])}** files modified",
+        f"- **{len(changes['deleted'])}** files deleted",
+        "",
+    ]
+
+
+def _build_footer(tmpl: _TemplateInfo) -> list[str]:
+    """Build the PR description footer with metadata.
+
+    Args:
+        tmpl: Template metadata container
+
+    Returns:
+        List of footer lines
+    """
+    lines = [
+        "---",
+        "",
+        "**🤖 Generated by [rhiza](https://github.com/jebel-quant/rhiza-cli)**",
+        "",
+    ]
+    if tmpl.repo:
+        lines.append(f"- Template: `{tmpl.repo}@{tmpl.branch}`")
+    if tmpl.last_sync:
+        lines.append(f"- Last sync: {tmpl.last_sync}")
+    lines.append(f"- Sync date: {datetime.now().astimezone().isoformat()}")
+    return lines
+
+
+def _generate_json_output(
+    changes: dict[str, list[str]],
+    categories: dict[str, list[str]],
+    tmpl: _TemplateInfo,
+) -> str:
+    """Generate a JSON representation of the change data.
+
+    Args:
+        changes: Dictionary of changes by type
+        categories: Files grouped by category
+        tmpl: Template metadata container
+
+    Returns:
+        JSON-formatted string
+    """
+    data = {
+        "template_repo": tmpl.repo,
+        "template_branch": tmpl.branch,
+        "last_sync": tmpl.last_sync,
+        "sync_date": datetime.now().astimezone().isoformat(),
+        "changes": changes,
+        "categories": categories,
+    }
+    return _json.dumps(data, indent=2)
+
+
+def _plain_file_section(lines: list[str], label: str, files: list[str]) -> None:
+    """Append a labelled block of files to *lines* in plain-text format.
+
+    Args:
+        lines: List to append lines to
+        label: Section label (e.g. "Added")
+        files: List of file paths
+    """
+    if not files:
+        return
+    lines.append(f"{label}:")
+    lines.extend(f"  {f}" for f in sorted(files))
+    lines.append("")
+
+
+def _generate_plain_output(
+    changes: dict[str, list[str]],
+    categories: dict[str, list[str]],
+    tmpl: _TemplateInfo,
+    options: SummariseOptions,
+) -> str:
+    """Generate plain-text output from change data.
+
+    Args:
+        changes: Dictionary of changes by type
+        categories: Files grouped by category
+        tmpl: Template metadata container
+        options: Output customisation options
+
+    Returns:
+        Plain-text formatted string
+    """
+    lines: list[str] = []
+
+    if options.include_header:
+        heading = options.title or "Template Synchronization"
+        lines.extend([heading, "=" * len(heading), ""])
+        if tmpl.repo:
+            lines.append(f"Template: {tmpl.repo}@{tmpl.branch}")
+            lines.append("")
+
+    total = sum(len(v) for v in changes.values())
+    if not total:
+        lines.append("No changes detected.")
+        return "\n".join(lines)
+
+    lines.append(
+        f"Changes: {len(changes['added'])} added, "
+        f"{len(changes['modified'])} modified, "
+        f"{len(changes['deleted'])} deleted",
+    )
+    lines.append("")
+
+    if options.include_categories:
+        for category, files in sorted(categories.items()):
+            lines.append(f"{category}:")
+            lines.extend(f"  {f}" for f in sorted(files))
+            lines.append("")
+    else:
+        for label, files in [
+            ("Added", changes["added"]),
+            ("Modified", changes["modified"]),
+            ("Deleted", changes["deleted"]),
+        ]:
+            _plain_file_section(lines, label, files)
+
+    if options.include_footer:
+        if tmpl.last_sync:
+            lines.append(f"Last sync: {tmpl.last_sync}")
+        lines.append(f"Sync date: {datetime.now().astimezone().isoformat()}")
+
+    return "\n".join(lines)
+
+
+def _generate_jinja2_output(template_path: Path, context: dict[str, Any]) -> str:
+    """Render output using a custom Jinja2 template file.
+
+    The *context* dict is passed directly to the template. It should contain at
+    minimum: ``template_repo``, ``template_branch``, ``last_sync``, ``sync_date``,
+    ``changes``, ``categories``, and ``title``.
+
+    Note:
+        Autoescape is disabled because this function generates plain text / Markdown,
+        not HTML.  Do **not** use the rendered output directly in a web context without
+        first escaping it, as the template content is not sanitised for HTML.
+
+    Args:
+        template_path: Path to the Jinja2 template file
+        context: Template context variables
+
+    Returns:
+        Rendered template string
+    """
+    template_text = template_path.read_text(encoding="utf-8")
+    env = jinja2.Environment(  # nosec B701
+        autoescape=False,  # noqa: S701
+        loader=jinja2.BaseLoader(),
+        keep_trailing_newline=True,
+    )
+    return env.from_string(template_text).render(**context)
+
+
+def _render_category_group(
+    lines: list[str],
+    categories: dict[str, list[str]],
+    changes: dict[str, list[str]],
+) -> None:
+    """Append the per-category collapsible markdown sections to *lines*.
+
+    Args:
+        lines: List to append lines to
+        categories: Files grouped by category
+        changes: Dictionary of changes by type, used to split each category's
+            files into added / modified / deleted subsections
+    """
+    lines.append("### 📁 Changes by Category")
+    lines.append("")
+
+    for category, files in sorted(categories.items()):
+        lines.append(f"#### {category}")
+        lines.append("")
+
+        category_added = [f for f in files if f in changes["added"]]
+        category_modified = [f for f in files if f in changes["modified"]]
+        category_deleted = [f for f in files if f in changes["deleted"]]
+
+        _add_category_section(lines, "Added", len(category_added), category_added, "✅")
+        _add_category_section(lines, "Modified", len(category_modified), category_modified, "📝")
+        _add_category_section(lines, "Deleted", len(category_deleted), category_deleted, "❌")
+
+
+def _render_flat_files(lines: list[str], changes: dict[str, list[str]]) -> None:
+    """Append a flat (uncategorised) markdown file listing to *lines*.
+
+    Args:
+        lines: List to append lines to
+        changes: Dictionary of changes by type
+    """
+    lines.append("### 📁 Changed Files")
+    lines.append("")
+    _add_category_section(lines, "Added", len(changes["added"]), changes["added"], "✅")
+    _add_category_section(lines, "Modified", len(changes["modified"]), changes["modified"], "📝")
+    _add_category_section(lines, "Deleted", len(changes["deleted"]), changes["deleted"], "❌")
+
+
+def _markdown_body(
+    changes: dict[str, list[str]],
+    categories: dict[str, list[str]],
+    tmpl: _TemplateInfo,
+    options: SummariseOptions,
+) -> str:
+    """Build the markdown PR description body.
+
+    Args:
+        changes: Dictionary of changes by type
+        categories: Files grouped by category
+        tmpl: Template metadata container
+        options: Output customisation options
+
+    Returns:
+        Markdown-formatted string
+    """
+    lines: list[str] = []
+
+    if options.include_header:
+        lines.extend(_build_header(tmpl.repo, title=options.title))
+
+    total_changes = sum(len(files) for files in changes.values())
+    if not total_changes:
+        lines.append("No changes detected.")
+        if options.include_footer:
+            lines.append("")
+            lines.extend(_build_footer(tmpl))
+        return "\n".join(lines)
+
+    lines.extend(_build_summary(changes))
+
+    if options.include_categories and categories:
+        _render_category_group(lines, categories, changes)
+    elif not options.include_categories:
+        _render_flat_files(lines, changes)
+
+    if options.include_footer:
+        lines.extend(_build_footer(tmpl))
+
+    return "\n".join(lines)
